Rearrange algorithms for modular arithmetic
